Test code for i2c is working
Communication with i2c between PIC32MX and MPU6050 is now working. This
test code is able to read from the MPU6050 address 0x75 (WHO_AM_I). The
value that is read is 0x68 which is the MPU6050 i2c device address. Also
a datasheet for PIC32MX Oscillators has been added.